improve details
