+++ src/components/layouts/DefaultLayout/Header/index.ts
@@ -1,0 +1,1 @@
+export { default as DefaultLayoutHeader } from "./Header";
