+++ src/containers/index.ts
@@ -1,4 +1,3 @@
-export * from "./layout";
 export * from "./about";
-export * from "./project";
 export * from "./contact";
+export * from "./project";
